requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
PyYAML>=6.0
python-dotenv>=0.20.0
twilio>=8.5.0
schedule>=1.1.0
pytest>=7.2.0
pytest-cov>=4.0.0
click>=8.1.0
redis>=4.5.0
Flask>=2.3.0
Werkzeug>=2.3.0
